requirements.txt
matplotlib
Pillow
numpy
scikit-learn
scipy
tqdm
